% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize}
\alias{normalize}
\title{Normalize gene expression matrix}
\usage{
normalize(gene)
}
\arguments{
\item{gene}{A gene expression matrix where each row is one gene and each column is one sample}
}
\value{
A gene expression matrix after nomalization. Genes that not express in every samples will be filtered.
}
\description{
Normalize the gene expression matrix using DESeq method. You could find detailed description in paper "Differential expression analysis for sequence count data"
}
\examples{
gene <- normalize(gene)
}
