% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\docType{data}
\name{test}
\alias{test}
\alias{gene}
\alias{snp}
\title{test: A test dataset for DEsingle}
\format{\itemize{
  \item gene. A matrix of raw read counts of scRNA-seq data which has 10 genes (rows) and 1529 cells (columns).
  \item snp A matrix of genotype of scRNA-seq data which has 20 variants (rows) and 1529 cells (columns).
}}
\usage{
data(test)
}
\description{
A toy dataset containing a single-cell RNA-seq (scRNA-seq) gene expression matrix and genotype matrix.
}
\details{
\itemize{
  \item gene. A matrix of raw read counts of scRNA-seq data which has 10 genes (rows) and 1529 cells (columns).
  \item snp A matrix of genotype of scRNA-seq data which has 20 variants (rows) and 1529 cells (columns).
}
}
\keyword{data}
